Implement null statements
Be careful, this is not the `null` keyword. It's a "null statement"
which means that it's just a statement that has only semicolon. This is
a valid syntax in C.
For example you can hava multiple semicolon statements like this:

```
;
;;;;
;
;;;;
```

This is not a very exciting feature on its own, but it will be needed
for the `for` loops.